requirements.txt
certifi==2022.9.24
# cffi>=1.15.1
# charset-normalizer>=3.0.0
chardet>=3.0.0
docopt>=0.6.2
ecolib>=2.05
idna==3.4
pip==22.3.1
pipreqs>=0.4.11
pycparser>=2.21
PyNaCl>=1.5.0
requests==2.28.1
setuptools>=65.5.1
urllib3==1.26.13
yarg>=0.1.9
beautifulsoup4>=4.11.1
redis>=4.3.4
websockets>=10.4